agent/requirements.txt
pydantic>=2.9.2
pymongo==4.10.1
mysql-connector-python

openai==1.47.0
tavily-python==0.5.0
kagglehub>=0.3.3

matplotlib==3.9.2
scipy==1.13.1
pandas==2.2.3
numpy==1.26.4

fastapi
uvicorn[standard]
